// frontend/src/api/generated/models/ticketLinks.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable ticket tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */
import type { TicketLinkRead } from './ticketLinkRead';

export interface TicketLinks {
  blocks?: TicketLinkRead[];
  blocked_by?: TicketLinkRead[];
  relates_to?: TicketLinkRead[];
  duplicates?: TicketLinkRead[];
  duplicated_by?: TicketLinkRead[];
}