// frontend/src/api/generated/models/ticketLinkRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable ticket tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */
import type { LinkedTicket } from './linkedTicket';

export interface TicketLinkRead {
  id: number;
  relation: string;
  ticket: LinkedTicket;
  created_at: string;
}